requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langchain_huggingface
llama-cpp-python
langserve
fastapi
uvicorn
sse_starlette
langchain-ollama